Update: fetchUtils/index ts로 변경
